Move header, footer and main to common components
